- restructured the tests back into the original "default" style
