--- test/test_main.dart
@@ -1,11 +1,0 @@
-import 'rendering/painter/all_painter_tests.dart';
-import 'widgets/label_builder/dynamic_timeline_label_builder_tests.dart';
-import 'widgets/dynamic_timeline_tests.dart';
-import 'widgets/timeline_item_tests.dart';
-
-void main(){
-  DynamicTimelineTests.run();
-  DynamicTimelineLabelBuilderTests.run();
-  TimelineItemTests.run();
-  AllPainterTests.run();
-}
